requirements.txt
# paddlepaddle-gpu
paddlepaddle
paddleocr
dataset
arrow
click
chardet
jieba
scikit-learn
opencv-python
pyautogui
redis
celery
pyyaml
numpy
eventlet